m_08_mongo/pyproject.toml
[tool.poetry]
name = "m_08_mongo"
version = "0.1.0"
description = ""
authors = ["Siracenco Serghei <siracencoserghei@gmail.com>"]
readme = "README.md"
packages = [{include = "m_08_mongo"}]

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.6.0"
mongoengine = "^0.27.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
